Level/Skills rework
I changed how levels are calculated, instead of setting the level, its determined by the entity's stats. This is more robust, because levels are relevant to the entity's difficulty and is standardized.

This required a lot to be modified within enemy and player class.

I rewrote a couple things within enemy to account for the change, as well as forced entities to have a constructor that gives it basic attributes.