Changing map infrastructure
Changed the way the map is interacted with and how it generates things. I also made a lot of broken things innaccessible for now so its not overwhelming with errors while testing.

I'm also thinking of moving all menus to a static class of their own, since all menus should do it manipulate data instead of created new stuff. Which also makes menus easier to change, as well as easier to access.